# server/openapi.yaml
openapi: 3.1.3
info:
  title: Blog API
  version: "1.0.0"
  description: |
    API backend pour un blog avec:
    - Auth par JWT access_token (cookie httpOnly)
    - Refresh token rotatif en base (table RefreshToken)
    - Gestion des utilisateurs, rôles, sécurité
    - CRUD des posts
    - Pagination standardisée
    - Récupération de mot de passe (reset sécurisé + rotation des sessions)
servers:
  - url: http://localhost:8080
    description: Dév local
...
components:
  securitySchemes:
    AccessTokenCookie:
      type: apiKey
      in: cookie
      name: access_token
      description: |
        Jeton d'accès JWT court.
        - Envoyé par le backend dans un cookie httpOnly `access_token`.
        - Le navigateur le renvoie automatiquement.
        - Dans Swagger/Postman, tu dois le simuler à la main.
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          properties:
            status:
              type: integer
              example: 401
            code:
              type: string
              example: UNAUTHORIZED
            message:
              type: string
              example: Non authentifié
            traceId:
              type: string
              description: ID de requête utile au debug
            stack:
              type: string
              description: Inclus seulement en dev
...
    UserPublic:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        name:
          type: string
          nullable: true
        role:
          type: string
          enum: [ROLE_USER, ROLE_ADMIN]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # /users/me renvoie la même forme publique
...
    RefreshBody:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          description: Refresh token courant
    RefreshResponse:
      type: object
      description: |
        Renvoie:
        - user
        - le NOUVEAU refreshToken
        - refreshExpiresAt
        Et met à jour le cookie httpOnly `access_token`.
      properties:
        user:
          $ref: "#/components/schemas/UserPublic"
        refreshToken:
          type: string
        refreshId:
          type: integer
        refreshExpiresAt:
          type: string
          format: date-time
...
paths:
...
  /auth/refresh:
    post:
      tags:
        - Auth
      summary: Rafraîchir la session (rotation du refresh token)
      description: |
        - Nécessite un refreshToken valide dans le body.
        - Renvoie un NOUVEAU refreshToken et remet un nouveau cookie access_token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshBody"
      responses:
        "200":
          description: Session régénérée
          headers:
            Set-Cookie:
              description: Nouveau cookie httpOnly `access_token`
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshResponse"
        "401":
          description: Refresh token invalide / expiré / révoqué
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...